+++ EmpiriaBMS.Models/Models/Other.cs
@@ -13,7 +13,7 @@ public class Other : Entity
     [Required]
     public string Name { get; set; }
 
-    public double MenHours { get; set; }
+    public long MenHours { get; set; }
 
     public float CompletionEstimation { get; set; }
